+++ packages/amplify-graphql-auth-transformer/src/resolvers/query.ts
@@ -47,6 +47,7 @@ import {
   RelationalPrimaryMapConfig,
   IDENTITY_CLAIM_DELIMITER,
 } from '../utils';
+import { aws_dynamodb as dynamodb } from 'aws-cdk-lib';
 
 const generateStaticRoleExpression = (roles: Array<RoleDefinition>): Array<Expression> => {
   const staticRoleExpression: Array<Expression> = [];
@@ -135,6 +136,7 @@ const generateAuthOnModelQueryExpression = (
   primaryFields: Array<string>,
   isIndexQuery = false,
   primaryKey: string | undefined = undefined,
+  sortKeyFields: Array<string>,
 ): Array<Expression> => {
   const modelQueryExpression: Expression[] = new Array<Expression>();
   const primaryRoles = roles.filter(r => primaryFields.includes(r.entity));
@@ -234,12 +236,28 @@ const generateAuthOnModelQueryExpression = (
             methodCall(ref('util.isNull'), ref('ctx.stash.authFilter')),
             not(ref('primaryFieldMap.isEmpty()')),
           ]),
+          /**
+           * The following code block is to handle the cases of dynamic claim field(owner/group)
+           * being part of the index sort key field and is missing from the query input arguments.
+           * If detected, the last item of the claim list will be added by default to the query condition with "eq".
+           * This restricts the records of other users returned from the query in such scenario
+           */
           compoundExpression([
+            set(ref('sortKeyFields'), list(sortKeyFields.map(sk => str(sk)))),
             forEach(ref('entry'), ref('primaryFieldMap.entrySet()'), [
-              qref(methodCall(ref('ctx.args.put'), ref('entry.key'), ref('entry.value'))),
+              ifElse(
+                methodCall(ref('sortKeyFields.contains'), ref('entry.key')),
+                compoundExpression([
+                  set(ref('entryVal'), ref('entry.value')),
+                  set(ref('lastIdx'), ref('entryVal.size() - 1')),
+                  set(ref('lastItem'), ref('entryVal.get($lastIdx)')),
+                  qref(methodCall(ref('ctx.args.put'),  ref('entry.key'), obj({ eq: ref('lastItem') }))),
+                ]),
+                qref(methodCall(ref('ctx.args.put'), ref('entry.key'), ref('entry.value'))),
+              ),
               set(ref(IS_AUTHORIZED_FLAG), bool(true)),
             ]),
-          ]),
+          ])
         ),
       );
     } else {
@@ -509,13 +527,15 @@ export const generateAuthExpressionForQueries = (
   providers: ConfiguredAuthProviders,
   roles: Array<RoleDefinition>,
   fields: ReadonlyArray<FieldDefinitionNode>,
-  primaryFields: Array<string>,
+  keySchema: dynamodb.CfnTable.KeySchemaProperty[],
   isIndexQuery = false,
   primaryKey: string | undefined = undefined,
 ): string => {
   const {
     cognitoStaticRoles, cognitoDynamicRoles, oidcStaticRoles, oidcDynamicRoles, apiKeyRoles, iamRoles, lambdaRoles,
   } = splitRoles(roles);
+  const primaryFields = keySchema.map((att) => att.attributeName);
+  const sortKeyFields = keySchema.filter((att) => att.keyType === 'RANGE').map((att) => att.attributeName);
   const getNonPrimaryFieldRoles = (
     rolesToFilter: RoleDefinition[],
   ): RoleDefinition[] => rolesToFilter.filter(role => !primaryFields.includes(role.entity));
@@ -540,7 +560,7 @@ export const generateAuthExpressionForQueries = (
         compoundExpression([
           ...generateStaticRoleExpression(cognitoStaticRoles),
           ...generateAuthFilter(getNonPrimaryFieldRoles(cognitoDynamicRoles), fields),
-          ...generateAuthOnModelQueryExpression(cognitoDynamicRoles, primaryFields, isIndexQuery, primaryKey),
+          ...generateAuthOnModelQueryExpression(cognitoDynamicRoles, primaryFields, isIndexQuery, primaryKey, sortKeyFields),
         ]),
       ),
     );
@@ -552,7 +572,7 @@ export const generateAuthExpressionForQueries = (
         compoundExpression([
           ...generateStaticRoleExpression(oidcStaticRoles),
           ...generateAuthFilter(getNonPrimaryFieldRoles(oidcDynamicRoles), fields),
-          ...generateAuthOnModelQueryExpression(oidcDynamicRoles, primaryFields, isIndexQuery, primaryKey),
+          ...generateAuthOnModelQueryExpression(oidcDynamicRoles, primaryFields, isIndexQuery, primaryKey, sortKeyFields),
         ]),
       ),
     );