+++ packages/graphql-elasticsearch-transformer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.54](https://github.com/aws-amplify/amplify-category-api/compare/graphql-elasticsearch-transformer@5.2.53...graphql-elasticsearch-transformer@5.2.54) (2023-05-17)
+
+**Note:** Version bump only for package graphql-elasticsearch-transformer
+
 ## [5.2.53](https://github.com/aws-amplify/amplify-category-api/compare/graphql-elasticsearch-transformer@5.2.52...graphql-elasticsearch-transformer@5.2.53) (2023-04-25)
 
 **Note:** Version bump only for package graphql-elasticsearch-transformer